requirements.txt
py-cord~=2.6.1
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
requests~=2.32.3
